добавил конвертеры (RoomConverters (для ROOM) и ExtendedIngredientCon…

+++ app/src/main/java/com/agobikk/cookeatenjoy/data/repository/RecipeRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.agobikk.cookeatenjoy.model.FoodInformation
 class RecipeRepositoryImpl {
     private var foodInformationRepo: List<FoodInformation> = emptyList()
 
-    fun getIngredientList(recipeId:Int):List<ExtendedIngredient>{
+    fun getIngredientList(recipeId:Long):List<ExtendedIngredient>{
       return foodInformationRepo.find { it.id == recipeId }?.extendedIngredient ?: emptyList()
     }
 }
